\name{Fisher}
\docType{data}
\alias{Fisher}
\title{Critical values of Fisher's distribution}
\description{This dataset gives the critical values for specific significance levels. To see this dataset you should write the following code: "read.csv("Fisher.csv")".}
\format{This dataset is a table that contains 38 observations and 4 variables.}
