\name{Simple_Regression-package}
\alias{Simple_Regression-package}
\alias{Simple_Regression}
\encoding{UTF-8}
\docType{package}
\title{Simple Linear Regression}

\description{This Package has been created to give to users an easy experience, in fact the functions that are contained in it are very easy to use and also their outputs are very clear. In the package there are two different functions that can be used in many fields: the first is a function that calculates a simple regression model, while the other is a function that calculates the F-Test about R-Squared (goodness of fit). In addition you can consult the package's documentation in which are contained more information regarding the way in which you can apply the functions and the way in which you can read (correctly) the outputs.}

\details{In this package there are two functions: SimpleLinearRegression(), F_test().}

\note{For further information, or if there is a malfunction, please contact the author through the following email address: \email{antoniop.r0033@gmail.com}}

\author{AntonioP-R}